Fixed type handling for boolean operators
Comparing values of different types except integer and real was somewhat
broken. Now it parses common.j to extract a type hierachy and uses the
lowest comman ancestor on that tree to insert the correct conv
instructions. With this change we also put the type hierachy into the
compiler state and write it to disk.