Promote integer literals to floats
When compiling Jass to Bytecode we now convert Int literals into Real
literals when the asked type is real.
This way `2*x` is the same as `2.0*x` (in the correct context) which we
need to have to be able to correctly match against the former case in the
rewrite step.