Make Hot.Instruction.Compiler.typeOfExpr monadic
This change is important to keep the bytecode correctly typed when
converting integer literals to reals in real context. Let's consider this
example: `R2S(-(180))`.
This would force the argument to R2S into real context but `typeOfExpr`
would correclty return `integer` for 180. So the emitted neg instruction
would convert from integer to the wanted real type.
The fix to this is to check the wanted type in typeOfExpr when an integer
literal is passed.