Handle null/code comparison
Previously `null` was always treated as of type handle in comparisons but
code is internally represented as integer so we had a mismatch.
To fix this we don't convert code to integer implicitly but we have to
call `code2int` explicitely. This makes sense since we can use code in
typechecking but int in compiling.