Fixed Jass Rewrite Engine
We had that `a-b-c` was rewritten to `a` based on the rewrite rule
`-(-a) --> a`. That was because we use zipWithM to match all arguments
to a potential call after checking if the name of the function matches.
That means if the name matches and all the arguments to the LHS of the
rule match the remaining arguments of the RHS got dropped (due to normal
zip behaviour). To fix this we simply check if the length of both
argument lists is equal.