Add warning if common.j hashes mismatch.
At compile time we use the Hashable instance for our Jass Ast and
store the hash. At init we take the first provided jass file and compute
hash over that files Ast aswell. Using an Ast hash will give use some
resiliance against stupid stuff like line endings but it is of course
not cryptographically secure. That shouldn't matter much as we only use
it as a potential error cause when a map doesn't start.