Generate a new JHCR file on newer patches.
On patches from 1.33 onward WC3 aggressively caches files called by Preloader.
To work around this we simply create a new file each time. For now we don't do
any shenanigans like trying to do non-predictable sequence numbers as i'm not
aware of any upsides but it might happen in the future if needed.