Use a random prefix for preload files
At `jhcr init` we generate a random prefix and store it inside the state file.
We do this because we want unique preload names. We need unique names because
once WarCraft 3 preloads an existing file it will cache that file and thus it
cannot be modified and then loaded again. But due to the previous naming scheme
we had very predictable names lying around. So the user could by accident load
totally unrelated changes into the map. While we were at it, we also added
checks so that all updates are loaded in order and trying to load an update
without there actually being one will not result in a broken state anymore.

Now we spam even more preload files than before so we also added an autoclean
flag to update.