Introduce new Opcode CCR
We have this rather obscure situation of changing the signature of a function
to and then using that function as a filterfunc in freshly loaded code. So far
this did not work. To make it work we introduced a new opcode name CCR meaning
change coderef which links an allready defined function id with a dummy id.
Then we check a table for these changes in Auto#_i2code and just use the newly
provided id in-place.